requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appscript==1.2.5
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
black==23.12.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
datasets==2.16.1
decorator==5.1.1
dill==0.3.7
distro==1.9.0
docopt==0.6.2
docx2pdf==0.1.8
duckdb==0.9.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2023.10.0
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.60.0
gspread==5.12.4
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.2.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.19
langchain-openai==0.0.5
langsmith==0.0.86
lxml==5.1.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.11.1
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
pathspec==0.12.1
pillow==10.2.0
pinecone-client==3.0.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
protobuf==3.20.3
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==3.17.4
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sql_metadata==2.10.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
st-gsheets-connection==0.0.3
streamlit==1.31.0
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
twilio==8.11.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
